+++ client/web/src/insights/pages/dashboards/dashboard-page/components/add-insight-modal/components/add-insight-modal-content/AddInsightModalContent.tsx
@@ -34,7 +34,11 @@ export const AddInsightModalContent: React.FunctionComponent<AddInsightModalCont
         onSubmit,
     })
 
-    const searchInput = useField('searchInput', formAPI)
+    const searchInput = useField({
+        name: 'searchInput',
+        formApi: formAPI,
+    })
+
     const {
         input: { isChecked, onChange, onBlur },
     } = useCheckboxes('insightIds', formAPI)