+++ client/web/src/insights/pages/insights/creation/lang-stats/utils/insight-sanitizer.ts
@@ -1,12 +1,13 @@
 import { camelCase } from 'lodash'
 
-import { InsightTypePrefix, LangStatsInsight } from '../../../../../core/types'
+import { InsightType, InsightTypePrefix, LangStatsInsight } from '../../../../../core/types'
 import { LangStatsCreationFormFields } from '../types'
 
 /**
  * Converter from creation UI form values to real insight object.
  * */
 export const getSanitizedLangStatsInsight = (values: LangStatsCreationFormFields): LangStatsInsight => ({
+    type: InsightType.Extension,
     // ID generated according to our naming insight convention
     // <Type of insight>.insight.<name of insight>
     id: `${InsightTypePrefix.langStats}.${camelCase(values.title)}`,